Moved NonHumanoidSetup class to its own cs file.
