+++ Editor/Scripts/Internal/NonHumanoidSetup.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1301cab760208c4489f65ad81924f43b
